Fix returned status code in case of partial data availability
Consider the case of downloading data for 6 months. This script does it month by
month. If data does not exist for the last month the return code was `-10`
before this fix, indicating that no data at all is available.

Now we only get a `-10` return code if indeed no data at all is available in the
requested period